% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC.R
\name{calc_auc}
\alias{calc_auc}
\title{Calculate the Area Under the Curve (AUC)}
\usage{
calc_auc(y, x, steps = NULL, norm_to_1 = TRUE)
}
\arguments{
\item{y}{Numeric vector of y-values.}

\item{x}{Numeric vector of x-values.}

\item{steps}{Optional. Number of steps to interpolate the data. If NULL, no interpolation is performed.}

\item{norm_to_1}{Logical. If TRUE (default), the calculated AUC is normalized to 1.}
}
\value{
The calculated Area Under the Curve (AUC).
}
\description{
This function calculates the Area Under the Curve (AUC) using the trapezoidal rule.
It can interpolate the input data to a specified number of steps and normalize the result to 1.
}
\examples{
# Basic usage
x <- c(0, 1, 2, 3, 4, 5)
y <- c(0, 1, 2, 3, 4, 5)
calc_auc(y, x)

# With interpolation
x <- c(0, 2, 4)
y <- c(0, 2, 4)
calc_auc(y, x, steps = 100)

# Without normalization
x <- c(0, 1, 2, 3, 4, 5)
y <- c(0, 1, 2, 3, 4, 5)
calc_auc(y, x, norm_to_1 = FALSE)

}
