% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_plate_data.R
\name{combine_summaries}
\alias{combine_summaries}
\title{Combine summary files within a single directory}
\usage{
combine_summaries(path, pattern = "_summary.csv", output_prefix = NULL)
}
\arguments{
\item{path}{A character string specifying the directory path to search for summary files.}

\item{pattern}{A character string specifying the pattern to match for summary files. Default is '_summary.csv'.}

\item{output_prefix}{A character string specifying the prefix to be added to the output file name. Default is NULL (no prefix).}
}
\description{
This function searches for files matching the specified pattern in the provided directory path,
reads the data from each file, adds a 'file' column with the base name of the file, and combines the data
into a single data.table. The combined data is then written to a file with the specified prefix and '_combined_summary.csv' suffix in the directory.
}
