% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_plate_data.R
\name{combine_summary_files}
\alias{combine_summary_files}
\title{Combine summary files from multiple directories}
\usage{
combine_summary_files(
  results_file_paths,
  pattern = "_summary.csv",
  output_prefix = NULL,
  parallel = FALSE,
  cores = NULL
)
}
\arguments{
\item{results_file_paths}{A character vector of directory paths to search for summary files.}

\item{pattern}{A character string specifying the pattern to match for summary files. Default is '_summary.csv'.}

\item{output_prefix}{A character string specifying the prefix to be added to the output file name. Default is NULL (no prefix).}

\item{parallel}{A logical value indicating whether to use parallel processing. Default is FALSE.}

\item{cores}{An integer specifying the number of cores to use for parallel processing. Default is NULL (automatically determined).}
}
\description{
This function applies the \code{combine_summaries} function to each directory path provided in \code{results_file_paths}.
}
\examples{
\dontrun{
results_dirs <- c("path/to/results1", "path/to/results2")
combine_summary_files(results_dirs, pattern = "_summary.csv", output_prefix = "experiment_1", parallel = TRUE, cores = 4)
}

}
