% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_report_func.R
\name{create_growth_plot}
\alias{create_growth_plot}
\title{Create Growth Plot}
\usage{
create_growth_plot(d)
}
\arguments{
\item{d}{A data.table containing the data to be plotted. It should have columns 'well', 'frame', 'area_pineg', 'area_pipos', 'row', and 'col'.}
}
\value{
A heatmap showing the growth approximation across the plate.
}
\description{
This function generates a heatmap providing a comprehensive overview of the growth patterns observed across all wells throughout the experiment.
The growth ratio is calculated for each well by computing the ratio of the maximum total area to the minimum total area across all time points.
The growth ratio is log-transformed to better visualize the wide range of growth patterns.
}
