% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_report_func.R
\name{create_object_class_plots}
\alias{create_object_class_plots}
\title{Create Object Class Distribution Plots}
\usage{
create_object_class_plots(d)
}
\arguments{
\item{d}{A data.table containing the data to be plotted. It should have columns 'frame', 'row', 'col', 'object_class', and 'total_count'.}
}
\value{
A grid of heatmaps showing the object class distribution at t0.
}
\description{
This function generates a grid of heatmaps visualizing the spatial distribution of object classes across the plate at the initial time point (t0).
The visualization aggregates data by summing the counts for each object class across all fields within each well.
}
