% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_report_func.R
\name{create_time_delay_plot}
\alias{create_time_delay_plot}
\title{Create PI Consistency Plot}
\usage{
create_time_delay_plot(d)
}
\arguments{
\item{d}{A data.table containing the data to be plotted. It should have columns 'file', 'well', 'frame', 'object_class', 'pi_class_pos', 'total_count', 'row', and 'col'.}
}
\value{
A heatmap showing the PI classification consistency across the plate.
}
\description{
This function generates a heatmap visualizing the consistency of Propidium Iodide (PI) classification across different wells and time points, focusing on single-cell objects.
For each well, frame, and object class, the fraction of PI-positive cells is calculated. The complement of the PI fraction represents the live cell fraction.
The standard deviation of the live fraction across all time points is computed for each well, serving as a measure of classification consistency over time.
}
