% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrapolate_t0.R
\name{extrapolate_t0_data}
\alias{extrapolate_t0_data}
\title{Extrapolate T0 Data}
\usage{
extrapolate_t0_data(d, min_t, cols)
}
\arguments{
\item{d}{A data.table containing the plate well data.}

\item{min_t}{A numeric value specifying the minimum timepoint threshold.}

\item{cols}{A character vector specifying the columns to be extrapolated.}
}
\value{
A data.table with extrapolated T0 data.
}
\description{
This function extrapolates the correct T0 given plate well data, where the time 0 is lagged due to
the fact that imaging starts at an earlier timepoint in the first wells than on the later wells.
It extrapolates the average values of the wells with timepoint < min_t to the rest of the plate.
}
