% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_report.R
\name{render_QC}
\alias{render_QC}
\title{Render the QC report}
\usage{
render_QC(path_to_data, report_name = NULL, template_file = NULL)
}
\arguments{
\item{path_to_data}{The path to the data file to be used in the report.}

\item{report_name}{The name of the report file. If not provided, the report will be saved with the name \code{path_to_data}_qc_report.pdf.}

\item{template_file}{The path to the R Markdown template file. If not provided, the default template in the package will be used.}
}
\description{
This function renders the QC report using the template in \code{R/qc_report_template.Rmd}.
}
