% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{theme_plateview}
\alias{theme_plateview}
\title{Plate View Theme for ggplot2}
\usage{
theme_plateview(
  col_var,
  row_var,
  col_lab = "Column",
  row_lab = "Row",
  fill_lab = NULL,
  title = NULL,
  subtitle = NULL
)
}
\arguments{
\item{col_var}{The variable representing the plate columns.}

\item{row_var}{The variable representing the plate rows.}

\item{col_lab}{The label for the x-axis (plate columns). Default is "Column".}

\item{row_lab}{The label for the y-axis (plate rows). Default is "Row".}

\item{fill_lab}{The label for the fill aesthetic. Default is NULL.}

\item{title}{The title of the plot. Default is NULL.}

\item{subtitle}{The subtitle of the plot. Default is NULL.}
}
\value{
A list of ggplot2 theme elements to be added to a ggplot object.
}
\description{
This function provides a consistent theme for plate view plots in ggplot2.
It sets the scales, labels, and theme elements for a plate view visualization.
}
\examples{
library(ggplot2)
# Create dummy data
data <- data.frame(
  row = rep(LETTERS[1:8], each = 12),
  col = rep(1:12, 8),
  value = runif(96)
)
ggplot(data, aes(x = col, y = row, fill = value)) +
  geom_tile() +
  theme_plateview(col_var = data$col, row_var = data$row,
                  fill_lab = "Value", title = "Plate View Plot")

}
